login, register and resend verification code is ok (issue #3).

+++ database/migrations/2020_12_11_165053_create_verify_codes_table.php
@@ -15,7 +15,7 @@ class CreateVerifyCodesTable extends Migration
     {
         Schema::create('verify_codes', function (Blueprint $table) {
             $table->id();
-            $table->integer('code');
+            $table->bigInteger('code');
             $table->boolean('used')->default(false);
             $table->bigInteger('employee_id')->unsigned();
